# src/main/java/com/gildong/gildongE/dto/NotificationRequest.java
package com.gildong.gildongE.dto;

public class NotificationRequest {
    private String userId;
    private String title;
    private String message;

    public String getUserId() {
        return userId;
    }

    public void setUserId(String userId) {
        this.userId = userId;
    }

    public String getTitle() {
        return title;
    }

    public void setTitle(String title) {
        this.title = title;
    }

    public String getMessage() {
        return message;
    }

    public void setMessage(String message) {
        this.message = message;
    }
}