# src/main/resources/static/openapi.yaml
openapi: 3.0.0
info:
  title: GildongE API
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local server
...
paths:
...
  /api/notifications:
    post:
      tags: [Notifications]
      summary: 알림 저장
      description: |
        - 클라이언트에서 생성된 알림을 서버 DB에 저장합니다.
        - 이후 사용자가 알림 목록 조회 시 포함됩니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationRequest"
      responses:
        '201':
          description: 알림 저장 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationResponse"
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/notifications/user/{userId}:
    get:
      tags: [Notifications]
      summary: 사용자별 알림 목록 조회
      description: |
        - 특정 사용자(userId)의 저장된 모든 알림을 생성 시각 역순으로 반환합니다.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: 알림 대상 사용자 고유 ID
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NotificationResponse"
        '404':
          description: 사용자 없음 또는 알림이 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: 에러 코드
        message:
          type: string
          description: 에러 메시지
...
    NotificationRequest:
      type: object
      required:
        - userId
        - category
        - title
        - message
      properties:
        userId:
          type: string
          description: 알림 대상 사용자 고유 ID
        category:
          type: string
          description: 알림 카테고리 - 예) 차량 점검, 안전, 차량 소모품
        title:
          type: string
          description: 알림 제목
        message:
          type: string
          description: 알림 메시지
    NotificationResponse:
      type: object
      properties:
        id:
          type: string
          description: 알림 고유 ID
        userId:
          type: string
          description: 알림 대상 사용자 고유 ID
        category:
          type: string
          description: 알림 카테고리 (차량 점검, 안전, 차량 소모품)
        title:
          type: string
          description: 알림 제목
        message:
          type: string
          description: 알림 메시지
        createdAt:
          type: string
          format: date-time
          description: 알림 생성 시각